SSMW-386 subsystems: tee: Fix TEE TA compilation warnings
Fixes TA warnings.

Fixes: 8335fded5104 ("SSMW-240 subsystem: tee: Create a common library")
Fixes: aa10810b3b1e ("SSMW-371 subsystems: tee: Implements MAC operation")

+++ core/subsystems/tee/lib_ta/core/common.c
@@ -12,7 +12,7 @@ TEE_Result check_operation_keys_usage(TEE_OperationHandle op,
 	TEE_Result res = TEE_SUCCESS;
 	TEE_OperationInfoMultiple *op_info = NULL;
 	TEE_OperationInfoKey *op_key_info = NULL;
-	uint32_t op_info_size;
+	size_t op_info_size;
 
 	if (!key_info && nb_keys) {
 		res = TEE_ERROR_BAD_PARAMETERS;
